+++ StRoot/TStarEventClass/TStarJetConstituent.cxx
@@ -12,7 +12,7 @@ TStarJetConstituent::TStarJetConstituent(){
 
 }
 
-TStarJetConstituent::TStarJetConstituent(double e, double px, double py, double pz){
+TStarJetConstituent::TStarJetConstituent(float e, float px, float py, float pz){
     _E = e;
     _Px = px;
     _Py = py;
@@ -30,8 +30,8 @@ TStarJetConstituent::~TStarJetConstituent(){
 
 }
 
-double TStarJetConstituent::Phi(){
-    double phi = Phi_Std();
+float TStarJetConstituent::Phi(){
+    float phi = Phi_Std();
     if(phi < 0.0) return phi+2.0*TMath::Pi();
     else if(phi > 2.0*TMath::Pi()) return phi-2.0*TMath::Pi();
     else return phi;
